Make documentation clear for non-privileged users
The previous snippets in the documentation only work if the user have
admin privileges. Documentation now makes more clear what queries need
to be executed for regular non-privileged users.

Also, documentation gives some clue on how to find out the exact value
of the cloud domain name to be passed as first argument to query.run()